requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
pyodbc
python-dotenv
pydantic
pydantic-settings